Tratativas de exclusão de imagem temporária. @#$

+++ app/src/main/java/com/namoadigital/prj001/ui/act071/Act071_Main.java
@@ -282,9 +282,10 @@ public class Act071_Main extends Base_Activity implements Act071_Main_Contract.I
                                     if (currentLength != previousLenght) {
                                         hasImageFileChanged = true;
                                     }
-                                    if(photo.exists()) {
+                                    if(photo.exists() && hasImageFileChanged) {
                                         copyFiles(ConstantBase.CACHE_PATH_PHOTO + "/" + TEMP_SUFIX_FILE + actionPhotoLocalPath,
                                                 ConstantBase.CACHE_PATH_PHOTO + "/" + actionPhotoLocalPath);
+                                        photo.delete();
                                     }else{
                                         if(mPresenter.fileExists(actionPhotoLocalPath)) {
                                             File originalPhoto = new File(ConstantBaseApp.CACHE_PATH_PHOTO + "/" + actionPhotoLocalPath);
@@ -357,9 +358,23 @@ public class Act071_Main extends Base_Activity implements Act071_Main_Contract.I
             if(mTicketCtrl.getAction().getAction_photo_local() == null){
                 mTicketCtrl.getAction().setAction_photo(null);
                 mTicketCtrl.getAction().setAction_photo_name(null);
+                deletePhotoFile(actionPhotoLocalPath);
             }
         }else {
             mTicketCtrl.getAction().setAction_photo_changed(0);
+        }
+    }
+
+    private void deletePhotoFile(String filename) {
+        if(mPresenter.fileExists(filename)){
+            try {
+                File file = new File(ConstantBase.CACHE_PATH_PHOTO + "/" + filename);
+                if(file != null && file.exists()) {
+                    file.delete();
+                }
+            }catch (Exception e ){
+                e.printStackTrace();
+            }
         }
     }
 
@@ -531,12 +546,12 @@ public class Act071_Main extends Base_Activity implements Act071_Main_Contract.I
     private void defineActionPhoto() {
         actionPhotoLocalPath = mPresenter.generateActionPhotoLocalPath(mTicketCtrl.getAction());
         //
-        final File sFile = new File(ConstantBase.CACHE_PATH_PHOTO + "/" + TEMP_SUFIX_FILE + actionPhotoLocalPath);
-        //
         if (mTicketCtrl.getAction().getAction_photo() == null && mTicketCtrl.getAction().getAction_photo_local() == null) {
-
-            if(sFile != null && sFile.exists()){
-                sFile.delete();
+            if(!bReadOnly) {
+                final File sFile = new File(ConstantBase.CACHE_PATH_PHOTO + "/" + TEMP_SUFIX_FILE + actionPhotoLocalPath);
+                if (sFile != null && sFile.exists()) {
+                    sFile.delete();
+                }
             }
             ivActionPhoto.setImageDrawable(getResources().getDrawable(android.R.drawable.ic_menu_camera));
         } else {
@@ -545,20 +560,23 @@ public class Act071_Main extends Base_Activity implements Act071_Main_Contract.I
                 //INICIAR SERVICE DOWNLOAD E CRIAR HANDLER PARA DE X em X SEGUNDOS VERIFICAR SE SERVIÇO PAROU DE RODAR E SE PAROU TENTA RESETAR IMAGE?
                 //CRIAR ASYNC_TAKS PARA DOWNLOAD?
                 //USAR GLIDE PARA BAIXAR A IMAGEM SETANDO ELA NO PATH DEFINITIVO? NECESSARIO ATUALIZAR O BANCO DEPOIS...
+                ivActionPhoto.setImageDrawable(getResources().getDrawable(R.drawable.sand_watch_transp));
                 Glide.with(context).asBitmap()
                     .load(mTicketCtrl.getAction().getAction_photo())
-                    .placeholder(R.drawable.sand_watch_transp)
                     .into(new CustomTarget<Bitmap>() {
                         @Override
                         public void onResourceReady(@NonNull Bitmap resource, @Nullable Transition<? super Bitmap> transition) {
                             ivActionPhoto.setImageBitmap(resource);
-                            try (FileOutputStream out = new FileOutputStream(sFile)) {
-                                resource.compress(Bitmap.CompressFormat.PNG, 100, out);
-                            } catch (IOException e) {
-                                e.printStackTrace();
+                            if(!bReadOnly) {
+                                final File sFile = new File(ConstantBase.CACHE_PATH_PHOTO + "/" + TEMP_SUFIX_FILE + actionPhotoLocalPath);
+                                try (FileOutputStream out = new FileOutputStream(sFile)) {
+                                    resource.compress(Bitmap.CompressFormat.PNG, 99, out);
+                                } catch (IOException e) {
+                                    e.printStackTrace();
+                                }
+                                previousLenght = sFile.length();
                             }
                         }
-
                         @Override
                         public void onLoadCleared(@Nullable Drawable placeholder) {
 
@@ -567,12 +585,19 @@ public class Act071_Main extends Base_Activity implements Act071_Main_Contract.I
 
             } else {
                 //Passa dados para arquivo temporario, mudanca feita para restauracao de info original
-                copyFiles(ConstantBase.CACHE_PATH_PHOTO + "/" + actionPhotoLocalPath,
-                        ConstantBase.CACHE_PATH_PHOTO + "/" + TEMP_SUFIX_FILE + actionPhotoLocalPath);
-                previousLenght = sFile.length();
-                String path = ConstantBaseApp.CACHE_PATH_PHOTO + "/" + TEMP_SUFIX_FILE + actionPhotoLocalPath;
-                Bitmap bitmap = BitmapFactory.decodeFile(path);
-                ivActionPhoto.setImageBitmap(bitmap);
+                if(!bReadOnly) {
+                    final File sFile = new File(ConstantBase.CACHE_PATH_PHOTO + "/" + TEMP_SUFIX_FILE + actionPhotoLocalPath);
+                    copyFiles(ConstantBase.CACHE_PATH_PHOTO + "/" + actionPhotoLocalPath,
+                            ConstantBase.CACHE_PATH_PHOTO + "/" + TEMP_SUFIX_FILE + actionPhotoLocalPath);
+                    previousLenght = sFile.length();
+                    String path = ConstantBaseApp.CACHE_PATH_PHOTO + "/" + TEMP_SUFIX_FILE + actionPhotoLocalPath;
+                    Bitmap bitmap = BitmapFactory.decodeFile(path);
+                    ivActionPhoto.setImageBitmap(bitmap);
+                }else{
+                    String path = ConstantBase.CACHE_PATH_PHOTO + "/" + actionPhotoLocalPath;
+                    Bitmap bitmap = BitmapFactory.decodeFile(path);
+                    ivActionPhoto.setImageBitmap(bitmap);
+                }
             }
         }
     }
@@ -654,7 +679,11 @@ public class Act071_Main extends Base_Activity implements Act071_Main_Contract.I
     }
 
     private void updateActionPhotoReference() {
-        setActinPhotoToView(TEMP_SUFIX_FILE + actionPhotoLocalPath);
+        if(bReadOnly) {
+            setActinPhotoToView(actionPhotoLocalPath);
+        }else{
+            setActinPhotoToView(TEMP_SUFIX_FILE + actionPhotoLocalPath);
+        }
     }
 
     @Override
@@ -677,7 +706,11 @@ public class Act071_Main extends Base_Activity implements Act071_Main_Contract.I
             Criacao de arquivo temporario para restaurar imagem original caso user saia da
             activity sem confirmar as alteracoes.
         */
-        bundle.putString(ConstantBase.PPATH, TEMP_SUFIX_FILE +actionPhotoLocalPath );
+        if(bReadOnly) {
+            bundle.putString(ConstantBase.PPATH, actionPhotoLocalPath);
+        }else {
+            bundle.putString(ConstantBase.PPATH, TEMP_SUFIX_FILE + actionPhotoLocalPath);
+        }
         bundle.putBoolean(ConstantBase.PEDIT, !bReadOnly);
         bundle.putBoolean(ConstantBase.PENABLED, !bReadOnly);
         bundle.putBoolean(ConstantBase.P_ALLOW_GALLERY, false);
@@ -819,6 +852,7 @@ public class Act071_Main extends Base_Activity implements Act071_Main_Contract.I
     @Override
     public void onBackPressed() {
         //super.onBackPressed();
+        deletePhotoFile(TEMP_SUFIX_FILE + actionPhotoLocalPath);
         mPresenter.onBackPressedClicked(requestingAct);
     }
